requirements.txt
dnspython>=2.4.2
python-dotenv>=1.2.1
requests>=2.32.5